requirements.txt
flask
flask-cors
requests
opencv-python
numpy
faiss-cpu
torch
transformers
pillow
